+++ packages/base/sync/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Fix `computeChanges` method to work with items that have null values
+
 ## [1.3.0] - 2025-04-24
 
 ### Added